+++ Zenith/EntityComponent/Components/Zenith_CameraComponent.cpp
@@ -131,9 +131,9 @@ void Zenith_CameraComponent::GetPosition(Zenith_Maths::Vector3& xOut)
 
 void Zenith_CameraComponent::GetFacingDir(Zenith_Maths::Vector3& xOut) const
 {
-	xOut.z = glm::cos(m_fYaw) * glm::cos(m_fPitch);
-	xOut.x = -glm::sin(m_fYaw) * glm::cos(m_fPitch);
-	xOut.y = glm::sin(m_fPitch);
+	xOut.z = static_cast<float>(glm::cos(m_fYaw) * glm::cos(m_fPitch));
+	xOut.x = static_cast<float>(-glm::sin(m_fYaw) * glm::cos(m_fPitch));
+	xOut.y = static_cast<float>(glm::sin(m_fPitch));
 	xOut = glm::normalize(xOut);  // Must assign result back
 }
